Fixed warning in the KissXML lib

+++ KeePassLib/DDXMLDocument+MKPAdditions.m
@@ -27,7 +27,7 @@
 	// Therefore, we call it again here just to be safe.
 	xmlKeepBlanksDefault(0);
 
-	xmlDocPtr doc = xmlReadIO(ioread, ioclose, ioctx, NULL, NULL, mask);
+	xmlDocPtr doc = xmlReadIO(ioread, ioclose, ioctx, NULL, NULL, (int)mask);
 	if (doc == NULL) {
 		if (error) {
             *error = [NSError errorWithDomain:@"DDXMLErrorDomain" code:1 userInfo:nil];
